Added Load Demo, Added Load/Save Functionality
Some parts of the code were refactored to allow a seamless save/load feature. As "config" already existed, it was slightly modified to include routes, and the "Load Demo" button formally in the code editor was abstracted to loading a demo project (LFO), now located under the help menu.

Audio context is now reset upon a config change.

Destination is now added automatically to routes.